Aula 4.14 aula 4.28 aula 5.2

+++ src/main/java/com/murillo/algafood/infra/repository/EstadoRepositoryImpl.java
@@ -1,15 +1,18 @@
 package com.murillo.algafood.infra.repository;
 
+import com.murillo.algafood.domain.exception.EntidadeNaoEncontradaException;
 import com.murillo.algafood.domain.model.Estado;
 import com.murillo.algafood.domain.repository.EstadoRepository;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
+import javax.transaction.Transactional;
 import java.util.List;
 
-@Component
+@Repository
 public class EstadoRepositoryImpl implements EstadoRepository {
 
     @PersistenceContext
@@ -24,18 +27,27 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 
     @Override
     public Estado buscar(Long id) {
-        return entityManager.find(Estado.class, id);
+        Estado estado = entityManager.find(Estado.class, id);
+        if (estado == null) {
+            throw new EntidadeNaoEncontradaException(String.format("Não existe estado com código %d", id));
+        }
+
+        return estado;
     }
 
     @Override
+    @Transactional
     public Estado salvar(Estado estado) {
         return entityManager.merge(estado);
     }
 
     @Override
-    public void remover(Estado estado) {
-        estado = entityManager.find(Estado.class, estado.getId());
+    @Transactional
+    public void remover(Long id) {
+        Estado estado = entityManager.find(Estado.class, id);
+        if (estado == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(estado);
-
     }
 }
